+++ src/main/java/xyz/srgnis/bodyhealthsystem/BHSMain.java
@@ -1,6 +1,7 @@
 package xyz.srgnis.bodyhealthsystem;
 
 import static net.minecraft.server.command.CommandManager.*;
+import static xyz.srgnis.bodyhealthsystem.network.ClientNetworking.handleHealthChange;
 
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
@@ -19,70 +20,26 @@ import org.slf4j.LoggerFactory;
 import xyz.srgnis.bodyhealthsystem.body.BodyPart;
 import xyz.srgnis.bodyhealthsystem.body.player.PlayerBodyProvider;
 import xyz.srgnis.bodyhealthsystem.client.hud.BHSHud;
+import xyz.srgnis.bodyhealthsystem.command.DevCommands;
+import xyz.srgnis.bodyhealthsystem.network.ClientNetworking;
+import xyz.srgnis.bodyhealthsystem.network.ServerNetworking;
 
 public class BHSMain implements ModInitializer {
 	// This logger is used to write text to the console and the log file.
 	// It is considered best practice to use your mod id as the logger's name.
 	// That way, it's clear which mod wrote info, warnings, and errors.
 
-	public static final String MOD_ID = "bodyhealthsystem";
+	public static final String MOD_ID = "bhs";
 	public  static Identifier MOD_IDENTIFIER = new Identifier(BHSMain.MOD_ID);
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
 	@Override
 	public void onInitialize() {
-		// This code runs as soon as Minecraft is in a mod-load-ready state.
-		// However, some things (like resources) may still be uninitialized.
-		// Proceed with mild caution.
 		HudRenderCallback.EVENT.register(new BHSHud());
-
-		//TODO: create a class for the network handlers
-		ClientPlayNetworking.registerGlobalReceiver(MOD_IDENTIFIER, (client, handler, buf, responseSender) -> {
-			while (buf.isReadable()) {
-				Identifier id = buf.readIdentifier();
-				float health = buf.readFloat();
-				float maxhealth = buf.readFloat();
-
-				client.execute(() -> {
-					// Everything in this lambda is run on the render thread
-					((PlayerBodyProvider) client.player).getBody().getPart(id).setHealth(health);
-				});
-			}
-		});
-
-		ServerPlayConnectionEvents.JOIN.register((serverPlayNetworkHandler,packetSender,server)->{
-			BHSMain.syncBody(serverPlayNetworkHandler.player);
-		});
-
-		CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> dispatcher.register(literal("bhs")
-				.executes(context -> {
-					context.getSource().getPlayer().sendMessage(Text.literal(((PlayerBodyProvider)context.getSource().getPlayer()).getBody().toString()));
-					return 1;
-				}).then(literal("reset").executes(context -> {
-					((PlayerBodyProvider)context.getSource().getPlayer()).getBody().healAll();
-					BHSMain.syncBody(context.getSource().getPlayer());
-					return 1;
-				})).then(literal("damage").executes(context -> {
-					((PlayerBodyProvider)context.getSource().getPlayer()).getBody().damage(1);
-					BHSMain.syncBody(context.getSource().getPlayer());
-					return 1;
-				})))
-		);
+		ClientNetworking.initialize();
+		ServerNetworking.initialize();
+		DevCommands.initialize();
 
 	}
-
-	//TODO: Create a class for this
-	public static void syncBody(PlayerEntity pe){
-		PacketByteBuf buf = PacketByteBufs.create();
-
-		for (BodyPart part : ((PlayerBodyProvider)pe).getBody().getParts()) {
-			buf.writeIdentifier(part.getIdentifier());
-			buf.writeFloat(part.getHealth());
-			buf.writeFloat(part.getMaxHealth());
-		}
-
-		ServerPlayNetworking.send( (ServerPlayerEntity) pe, BHSMain.MOD_IDENTIFIER, buf);
-	}
-
 
 }
