+++ src/main/java/xyz/srgnis/bodyhealthsystem/mixin/OnDamageMixin.java
@@ -12,6 +12,7 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 import xyz.srgnis.bodyhealthsystem.BHSMain;
 import xyz.srgnis.bodyhealthsystem.body.player.PlayerBodyProvider;
 import xyz.srgnis.bodyhealthsystem.body.player.PlayerBody;
+import xyz.srgnis.bodyhealthsystem.network.ServerNetworking;
 
 @Mixin(PlayerEntity.class)
 public abstract class OnDamageMixin extends LivingEntity {
@@ -26,7 +27,7 @@ public abstract class OnDamageMixin extends LivingEntity {
         if(livingEntity instanceof PlayerEntity){
             PlayerBody body = ((PlayerBodyProvider)livingEntity).getBody();
             body.damage(amount);
-            BHSMain.syncBody((PlayerEntity)livingEntity);
+            ServerNetworking.syncBody((PlayerEntity)livingEntity);
         }
 
         return false;
